Replace Spiral with Mixture
Replaced the spiral function with a mixture of the wave and polynomial functions.
This is because the spiral was a difficult fit, and was more a test of data cleaning than of the model. The new mixture model is a more 'realistic' function, compared to the sorts of data we are likely to see in real datasets.